mac-requirements.txt
asgiref==3.5.0
async-timeout==4.0.2
attr==0.3.1
Babel==2.9.1
Deprecated==1.2.13
Django==3.2
django-appconf==1.0.5
django-colorfield==0.6.3
django-mathfilters==1.0.0
django-phonenumber-field==6.1.0
django-redis==5.2.0
django-select2==7.10.0
django-simple-history==3.0.0
packaging==21.3
phonenumbers==8.12.45
Pillow==9.0.1
pyparsing==3.0.7
pytz==2022.1
redis==4.2.0
SQLAlchemy==1.4.32
sqlparse==0.4.2
typing_extensions==4.1.1
wrapt==1.14.0
